+++ src/main/java/com/ontologycentral/osmwrap/webapp/POIServlet.java
@@ -63,6 +63,8 @@ public class POIServlet extends HttpServlet {
 				return;
 			}
 
+			resp.setHeader("X-Upstream-Source", archive + "?" + postData);
+
 			InputStream is = response.body();
 
 			if (wantJson) {